% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childcare.R
\docType{data}
\name{childcare}
\alias{childcare}
\title{Childcare worker vaccination data}
\format{A data frame with 341 rows and 4 variables:
 \describe{
  \item{Flu_5yr}{Whether respondent received flu vaccine in last 5 years}
  \item{HepA}{Whether respondent received full series of Hepatitis A vaccine}
  \item{Pertussis_10yr}{Whether respondent received pertussis booster in last
  10 years}
  \item{Flu_last_year}{Whether respondent received flu vaccine during last 
  flu season}
 }}
\source{
Data comes from study described in Rebmann, Terri, Travis M Loux, and
  Brendan Daly. "Uptake and Beliefs Regarding Seasonal Influenza Vaccination
  Among Childcare Center Staff, Administrators, and Parents." Infection
  Control & Hospital Epidemiology 38, no. 03 (March 2017): 367-370.
  \url{https://doi.org/10.1017/ice.2016.263}.
}
\usage{
childcare
}
\description{
A dataset containing childcare worker vaccination status for various
vaccinations. All variables are coded as 1 = Yes, 2 = No, 3 = Don't Know.
}
\keyword{datasets}
