% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civilwar.R
\docType{data}
\name{civilwar}
\alias{civilwar}
\title{Relevance of the Civil War}
\format{A data frame with 2168 rows and 2 variables:
 \describe{
  \item{id}{Participant identification number}
  \item{relevant}{Whether respondent believed Civil War is still relevant}
 }}
\source{
Data set created from the summarized data in
  \url{https://www.openintro.org/go?id=isrs1&referrer=/book/isrs/index.php}
  (pp 153).
}
\usage{
civilwar
}
\description{
Responses to a survey asking US adults whether they thought the Civil War
was still relevant to American politics and political life.
}
\keyword{datasets}
