% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatechange.R
\docType{data}
\name{climatechange}
\alias{climatechange}
\title{Act on climate change}
\format{
A data frame with 13664 rows and 2 variables:
 \describe{
  \item{generation}{Respondent's generation cohort}
  \item{action}{Whether respondent reported taking action to 
  address climate change in last year}
 }
}
\source{
Data set created from the summarized data in 2nd edition of
  \url{https://www.openintro.org/book/ims/} (pp 311).
}
\usage{
climatechange
}
\description{
Data from a study of Pew Research poll asking respondents whether 
they have personally taken action to help address climate change 
within the last year.
}
\keyword{datasets}
