% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flint.R
\docType{data}
\name{flint}
\alias{flint}
\title{Flint, Michigan, lead levels}
\format{A data frame with 71 rows and 3 variables: 
 \describe{
  \item{sample}{Sample identification number}
  \item{lead}{The recorded lead concentration in ppb (parts per billion)}
  \item{lead2}{The sample reported to the EPA, with removed samples labelled 
  as \code{NA}}
 }}
\source{
\url{file:///C:/Users/loux/Zotero/storage/MWUIW9I6/all-flint-water-crisis-related-documents-concerning-mdeq-and-mdhhs-obtained-so-far-via-freedom-.html}
}
\usage{
flint
}
\description{
A dataset containing lead levels from the Michigan Department of
Environmental Quality (MDEQ) study on lead levels in Flint, MI, drinking
water after the city switched water sources to the Flint River.
}
\keyword{datasets}
