% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flint_vt.R
\docType{data}
\name{flint_vt}
\alias{flint_vt}
\title{Flint, Michigan, lead levels as recorded by Virginia Tech study team}
\format{
A data frame with 271 rows and 1 variable: 
 \describe{
  \item{Pb_ppb}{The recorded lead concentration in ppb (parts per billion)}
 }
}
\source{
\url{http://flintwaterstudy.org/2015/12/complete-dataset-lead-results-in-tap-water-for-271-flint-samples/}
}
\usage{
flint_vt
}
\description{
A dataset containing lead levels from the Virginia Tech citizen science study
on lead levels in Flint, MI, drinking water. This data set was used to
counter the original study which purported to show that residential lead
levels were safe for drinking.
}
\keyword{datasets}
