% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss.R
\docType{data}
\name{gss}
\alias{gss}
\title{Marijuana legalization}
\format{A data frame with 1259 rows and 2 variables:
 \describe{
  \item{id}{Participant identification number}
  \item{legalize}{Respondent's answer to question}
 }}
\source{
Data set created from the summarized data in
  \url{https://www.openintro.org/go?id=isrs1&referrer=/book/isrs/index.php}
  (pp 153).
}
\usage{
gss
}
\description{
Geneal Social Survey results from a question asking "Do you think the use of
marijuana should be made legal, or not?"
}
\keyword{datasets}
