% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\docType{data}
\name{health}
\alias{health}
\title{General health measurements from a sample of research participants}
\format{A data frame with 80 rows and 14 variables: 
 \describe{
  \item{gender}{Participant gender \code{0} for male, \code{1} for female}
  \item{age}{Age in years}
  \item{ht}{Height in inches}
  \item{wt}{Weight in pounds}
  \item{waist}{Waist circumference in cm}
  \item{pulse}{Resting heart rate in beats per minute}
  \item{sys}{Systolic blood pressure}
  \item{dias}{Diastolic blood pressure}
  \item{chol}{Blood cholesterol}
  \item{bmi}{Body mass index}
  \item{leg}{idk}
  \item{elbow}{Elbow circumference in inches}
  \item{wrist}{Wrist circumference in inches}
  \item{arm}{idk}
 }}
\source{
\url{https://www.pearson.com/us/higher-education/product/Triola-Biostatistics-for-the-Biological-and-Health-Sciences-with-Statdisk/9780321194367.html}
}
\usage{
health
}
\description{
A dataset containing measurements such as height, weight, waist 
circumference, blood pressure, and cholesterol.
}
\keyword{datasets}
