% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv.R
\docType{data}
\name{hiv}
\alias{hiv}
\title{HIV is Sub-Saharan Africa}
\format{
A data frame with 240 rows and 2 variables:
\describe{
  \item{drug}{Drug assigned to study participant}
  \item{failure}{Whether the participant experienced
  virologic failure (HIV infection became worse) at
  24-week follow-up}
}
}
\source{
Data set created from the summarized data in 2nd edition of
  \url{https://www.openintro.org/book/ims/} (pp 297).
}
\usage{
hiv
}
\description{
Data from a randomized trial of viral suppression medication
among women with HIV in sub-Saharan Africa. The study
compared two viral suppressant drugs, Nevaripine and
Lopinavir, given to women after successful child-birth.
}
\keyword{datasets}
