% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideology.R
\docType{data}
\name{ideology}
\alias{ideology}
\title{Political ideology}
\format{A data frame with 835 rows and 2 variables: 
 \describe{
  \item{gender}{Participant gender}
  \item{ideology}{Self-reported political ideology. Coded \code{1} through 
  \code{5} for very liberal, somewhat liberal,  moderate, somewhat 
  conservative, and very conservative, repsectively}
 }}
\source{
unknown
}
\usage{
ideology
}
\description{
A dataset containing gender and political ideology of 835 study participants.
}
\keyword{datasets}
