% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immigrant.R
\docType{data}
\name{immigrant}
\alias{immigrant}
\title{Fictional data set on immigrant depression}
\format{A data frame with 275 rows and 2 variables: 
 \describe{
  \item{id}{Study participant identification number}
  \item{depr}{Indicator variable for diagnosis of clinical depression}
 }}
\usage{
immigrant
}
\description{
A dataset containing clinical depression diagnosis for 275 U.S. immigrants.
}
\keyword{datasets}
