% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhp.R
\docType{data}
\name{nhp}
\alias{nhp}
\title{National health plan}
\format{A data frame with 1262 rows and 2 variables: 
 \describe{
  \item{political}{Political party repondent identified with} 
  \item{support}{Whether respondent supported a national health plan}
 }}
\source{
Data set created from the summarized data in 4th edition of
  \url{https://www.openintro.org/stat/textbook.php} (pp 216).
}
\usage{
nhp
}
\description{
Responses to a Kaiser Family Foundation poll asking US adults whether they
supported a national health plan.
}
\keyword{datasets}
