% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offshore.R
\docType{data}
\name{offshore}
\alias{offshore}
\title{SUpport for offshore dirlling}
\format{A data frame with 827 rows and 2 variables:
 \describe{
  \item{grad}{Whether the respondent has an undergraduate college degree}
  \item{drill}{Respondent's support for offshore drilling}
 }}
\usage{
offshore
}
\description{
A dataset containing data about respondents' educational attainment and
support for oil drilling off the coast of California.
}
\keyword{datasets}
