% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.R
\docType{data}
\name{privacy}
\alias{privacy}
\title{Facebook privacy settings}
\format{A data frame with 806 rows and 2 variables:
 \describe{
  \item{gender}{Respondent's gender}
  \item{adjust}{Respondent's answer to question}
 }}
\source{
Data set created from the summarized data in
  \url{https://www.openintro.org/go?id=isrs1&referrer=/book/isrs/index.php}
  (pp 162).
}
\usage{
privacy
}
\description{
Responses to a survey question asking adult Facebook users "Do you know
how to adjust your Facebook privacy settings to control what people can
and cannot see?"
}
\keyword{datasets}
