% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quitters.R
\docType{data}
\name{quitters}
\alias{quitters}
\title{Smoking cessation}
\format{
A data frame with 300 rows and 2 variables:
 \describe{
  \item{treatment}{Participant's treatment group: \code{1} for treatment, 
  \code{0} for control}
  \item{quit}{Whether participant quit smoking by end of study: \code{1} for 
  yes, \code{0} for no}
 }
}
\source{
Data set created from the summarized data in 4th edition of
  \url{https://www.openintro.org/stat/textbook.php} (pp 245).
}
\usage{
quitters
}
\description{
Data from a study of a smoking cessation study. Participants in the treatment
group used a nicotine patch and met regularly with a support group. The
Control group wore patches but did not meet with a support group.
}
\keyword{datasets}
