% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\docType{data}
\name{sleep}
\alias{sleep}
\title{New York residents sleep}
\format{A data frame with 25 rows and 3 variables: 
 \describe{
  \item{id}{Participant identification number}
  \item{gender}{Participant gender}
  \item{hour}{Average number of hours slept}
 }}
\source{
Data set created from the summarized data in 
  \url{https://www.openintro.org/book/isrs/} (pp 200).
}
\usage{
sleep
}
\description{
A dataset containing gender and average hours of sleep for 25 NYC residents.
}
\keyword{datasets}
