% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallpox.R
\docType{data}
\name{smallpox}
\alias{smallpox}
\title{Smallpox}
\format{A data frame with 6224 rows and 2 variables:
 \describe{
  \item{inoculated}{Whether the individual was inoculated with smallpox}
  \item{result}{Whether the person lived or died}
 }}
\source{
Diez, D. M., Cetinkaya-Rundel, M., & Barr, C. D., (2019). *OpenIntro
statistics* (pp. 99). OpenIntro.
}
\usage{
smallpox
}
\description{
A dataset containing treatment and outcome variables of patients from a
smallpox innoculation study held in Boston, MA, in 1721.
}
\keyword{datasets}
