% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_depression.R
\docType{data}
\name{state_depression}
\alias{state_depression}
\title{Depression rates by race and state of residence}
\format{
A data frame with 51 rows and 3 variables: 
 \describe{
  \item{depression_w}{Depression rate for white residents}
  \item{depression_b}{Depression rate for Black residents}
  \item{depression_h}{Depression rate for Hispanic residents}
 }
}
\source{
Created from 2015 Behavioral Risk Factor Surveillance System public
data \url{https://www.cdc.gov/brfss/annual_data/annual_2015.html}.
}
\usage{
state_depression
}
\description{
A dataset containing depression rates for white, Black, and Hispanic 
residents of the 50 United States and the District of Columbia. State is 
provided as row name.
}
\keyword{datasets}
