% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\docType{data}
\name{students}
\alias{students}
\title{Fictional data set counting emergency department visits}
\format{
A data frame with 20 rows and 2 variables: 
 \describe{
  \item{id}{Student study identification number} 
  \item{er}{Number of emergency department visits in the previous academic 
  year} 
 }
}
\usage{
students
}
\description{
A dataset containing number of visits to teh emergency department for each of
20 students during the school year.
}
\keyword{datasets}
