% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYG.R
\docType{data}
\name{SYG}
\alias{SYG}
\title{Florida Stand Your Ground court cases}
\format{
A data frame with 220 rows and 3 variables: 
 \describe{
  \item{Defendant}{Race of the defendant} 
  \item{Decision}{Jury decision in the charge of the case, usually murder or 
  manslaughter} 
  \item{Victim}{Race of the victim}
 }
}
\source{
Data set created from the tabulated data available at
\url{https://doi.org/10.1080/10691898.2015.11889739}.
}
\usage{
SYG
}
\description{
A dataset containing information about homicide trials in the state of 
Florida in which the defendant used Florida's "Stand Your Ground" defense.
}
\keyword{datasets}
