% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss.R
\docType{data}
\name{gss}
\alias{gss}
\title{Marijuana legalization}
\format{A data frame with 1578 rows and 2 variables: 
 \describe{
  \item{id}{Participant identification number}
  \item{legalize}{Respondent's response to question}
 }}
\source{
Data set created from the summarized data in 4th edition of
  \url{https://www.openintro.org/stat/textbook.php} (pp 216).
}
\usage{
gss
}
\description{
Geneal Social Survey results from a question asking "Do you think the use of
marijuana should be made legal, or not?"
}
\keyword{datasets}
