% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhp.R
\docType{data}
\name{nhp}
\alias{nhp}
\title{National health plan}
\format{A data frame with 2168 rows and 2 variables:
 \describe{
  \item{political}{Political party repondent identified with}
  \item{support}{Whether respondent supported a national health plan}
 }}
\source{
Data set created from the summarized data in
  \url{https://www.openintro.org/go?id=isrs1&referrer=/book/isrs/index.php}
  (pp 153).
}
\usage{
nhp
}
\description{
Responses to a survey asking whether US adults supported a
national health plan.
}
\keyword{datasets}
